+++ mixbookBackend/Mixbook-java/src/main/java/com/mixbook/springmvc/Services/RecipeService.java
@@ -158,6 +158,6 @@ public interface RecipeService {
 	 * @param difficulty the recipe's difficulty to validate.
 	 * @return true if a recipe's difficulty is valid, or false if it is not valid.
 	 */
-	boolean isRecipeDifficultyValid(int difficulty);
+	boolean isRecipeDifficultyValid(byte difficulty);
 
 }
